refactor: extract ScoopApprovalRouter and ScoopIdleTimers (refs #1091)
Two more single-responsibility collaborators pulled out of Orchestrator.

ScoopApprovalRouter owns the cone-mediated sudo-request lifecycle: the
ConeRequestRegistry, the enqueue/deliver/persist flow, the per-scoop
SudoBroker factory, and the lick-card flip-on-resolve. The fail-closed
contract is preserved verbatim (no cone / unknown scoop / delivery
failure / unregister / timeout all resolve `deny`).

ScoopIdleTimers owns the per-scoop "no work received yet" notifier: the
Map<jid, Timeout>, the start/clear/clearAll surface, and the fire-once
cone notification. SCOOP_IDLE_TIMEOUT_MS is re-exported from the new
module so external imports keep working unchanged.

orchestrator.ts: 2129 -> 1855 lines (-274 this round; -1245 from the
original ~3,100). All 92 scoops/orchestrator tests stay green.
